requirements.txt
numpy==1.26.4
scikit-surprise==1.1.3
scipy==1.12.0
